requirements.txt
Flask==3.0.3
pytrends==4.9.2
supabase==2.5.0
python-dotenv==1.0.1
